Recharging drone and management of sensor data - WIP

+++ src/main/java/threads/DronesInput.java
@@ -5,7 +5,10 @@ import com.sun.jersey.api.client.Client;
 import com.sun.jersey.api.client.ClientHandlerException;
 import com.sun.jersey.api.client.WebResource;
 import modules.Drone;
+
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
 
 /*
  *
@@ -23,27 +26,38 @@ public class DronesInput extends Thread{
     }
 
     public void run() {
-
-        System.out.println("> Hit return to stop...");
-
-        try {
-            System.in.read();
-            Client client = Client.create();
-
-            System.out.println("> Removing the drone from the network...");
-            //Remove a drone to the REST Server
-            String deletePath = "/drones/"+d.getId();
-            WebResource webResource = client.resource(RESTServerAddress+deletePath);
+        while(true) {
+            System.out.println("> Digit 'exit' to stop...");
+            System.out.println("> Digit 'recharge' to recharge the drone's battery...");
             try {
-                webResource.type("application/json").delete();
+                InputStreamReader streamReader = new InputStreamReader(System.in);
+                BufferedReader reader = new BufferedReader(streamReader);
 
-            } catch (ClientHandlerException e) {
-                System.out.println("> Server not reachable");
+                String choice = reader.readLine();
+
+                if (choice.equals("exit")) {
+                    Client client = Client.create();
+                    System.out.println("> Removing the drone from the network...");
+                    //Remove a drone to the REST Server
+                    String deletePath = "/drones/" + d.getId();
+                    WebResource webResource = client.resource(RESTServerAddress + deletePath);
+                    try {
+                        webResource.type("application/json").delete();
+
+                    } catch (ClientHandlerException e) {
+                        System.out.println("> Server not reachable");
+                    }
+                    System.out.println("> Drone removed...");
+                    System.exit(0);
+                } else if (choice.equals("recharge")) {
+                    RechargingThread rechargingThread = new RechargingThread(d);
+                    rechargingThread.run();
+                } else {
+                    System.out.println("> Unknown command");
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
             }
-            System.out.println("> Drone removed...");
-            System.exit(0);
-        } catch (IOException e) {
-            e.printStackTrace();
         }
     }
 }
